00029 - Refactor EditableFieldFrame and related components into frames for improved structure; remove unused files and enhance layout management with EditablePartFrame and EditablePartFrameBase.
