+++ src/components/common/inputs/frames/EditablePartFrame.tsx
@@ -1,9 +1,5 @@
 import React, { useState } from "react";
-import EditButton from "../buttons/EditButton";
-import OkButton from "../buttons/OkButton";
-import CancelButton from "../buttons/CancelButton";
-import InputLabel from "./InputLabel";
-import { getHeightBySize, SizeType } from "../sizeUtils";
+import { getHeightBySize, SizeType } from "../../sizeUtils";
 import EditablePartFrameBase from "./EditablePartFrameBase";
 
 export interface EditablePartFrameChildProps {
